Move http package to internals
It's not ready for external exposure
